Add sensors to NUgus model (#9)
* Add camera devices to model

* Update NUgus.proto

* Rotate the cameras to face correct direction

* Remove noise and antiAliasing

* Add accelerometer and gyro

Positioned at [0,0,0.2] from torso space origin.

* Added Accelerometer & Gyro Placeholders

Added placeholders for the required parameters of both sensor objects. Current values are not correct and need to be edited.

* Accel Params Complete

The LUT table and the resolution parameters have been implemented accordingly to the ISS331dlh accelerometer.

* Gyroscope Params Filled

Completed the LUT list for the Gyroscope and the resolution value. Assumed sample rate was 400Hz in order to get the noise values.

* Added Lens and Focus Nodes to Cameras

Added nodes to camera objects for  convenience in future development.

* Added helpful comments and fixed Bug

Added comments describing the calculation Process of the noise values and assumptions made. In addition it was found that the first LUT values were incorrect and were corrected in this commit.

* Revert "Added helpful comments and fixed Bug"

This reverts commit 90847496dcddfe9841e96aa997d653684c27923b.

* Remove vs code folders and fix some formatting

* Add position sensor resolution

Sourced from the model specifications document

* Improved LUT

I have modified the LUT tables such that the RV's have a simulated 'DC gain' applied to them. This is advantageous as we work up from zero, as opposed to working up from -x to +x, as there will be no asymptote with the noise%

* Fixed Resolution Mismatch

The data calculated within the report did not match the data within the .PROTO file, hence it was updated.

* Update .gitignore

* Revert gitignore

* Update camera

* Add max focal distance

* Add reasonable noise and motionblur values

these are just guessed, they look sort of ok

* Adjust motion blur

and formatting

* Minimal noise

* Believe I've merged 'main' into 'Sensors'

* Revert "Believe I've merged 'main' into 'Sensors'"

This reverts commit fc8fd9ac96cc810f6ec6aa2ca6ea4048e012e9ca.

* Fix LUT indentation

* Remove camera parameters

* Remove LUT comments and add model docs

* Add comments

Co-authored-by: Benjamin Young <benjaminyoung7198@gmail.com>
Co-authored-by: Ysobel <ysobelsims@gmail.com>