Add motor parameters (#22)
* Add PID control gains for motors

PID gains sourced from Simscape simulation

* Add camera devices to model

* Update NUgus.proto

* Rotate the cameras to face correct direction

* Remove noise and antiAliasing

* Add accelerometer and gyro

Positioned at [0,0,0.2] from torso space origin.

* Added Accelerometer & Gyro Placeholders

Added placeholders for the required parameters of both sensor objects. Current values are not correct and need to be edited.

* Add motor params

- Adds constants for the MX106 and MX64 parameters and assigns to correct joints

- Sets selfCollision to TRUE

* Accel Params Complete

The LUT table and the resolution parameters have been implemented accordingly to the ISS331dlh accelerometer.

* Gyroscope Params Filled

Completed the LUT list for the Gyroscope and the resolution value. Assumed sample rate was 400Hz in order to get the noise values.

* Added Lens and Focus Nodes to Cameras

Added nodes to camera objects for  convenience in future development.

* Added helpful comments and fixed Bug

Added comments describing the calculation Process of the noise values and assumptions made. In addition it was found that the first LUT values were incorrect and were corrected in this commit.

* Revert "Added helpful comments and fixed Bug"

This reverts commit 90847496dcddfe9841e96aa997d653684c27923b.

* Remove vs code folders and fix some formatting

* Add position sensor resolution

Sourced from the model specifications document

* Add backlash to joints

- Causing segmentation fault on 2021b (robocup branch)

* Improved LUT

I have modified the LUT tables such that the RV's have a simulated 'DC gain' applied to them. This is advantageous as we work up from zero, as opposed to working up from -x to +x, as there will be no asymptote with the noise%

* Add simplified collision boxes

* Fixed Resolution Mismatch

The data calculated within the report did not match the data within the .PROTO file, hence it was updated.

* Update .gitignore

* Fix typo

* Update damping

original doc was wrong

* Newline

* Revert gitignore

* Update camera

* Add max focal distance

* Add reasonable noise and motionblur values

these are just guessed, they look sort of ok

* Adjust motion blur

and formatting

* Remove segmentation

* Minimal noise

* Fix backlash issues

Add motor parameters to left_ankle_pitch

* Remove duplicate variables

* Reduce motion blur

* Expose camera parameters

* Add missing backlash parameters from joints

* Remove unneeded camera params

* Remove unneeded bounding boxes

* Remove PID parameter

This will be overridden when servos targets are sent

* Add missing backlash param

* Remove textures

* Fix indentation on comments

Co-authored-by: Benjamin Young <benjaminyoung7198@gmail.com>
Co-authored-by: Ysobel <ysobelsims@gmail.com>