Adds receiving ActuatorRequests and sending SensorMeasurements (#31)
* Add camera devices to model

* Update NUgus.proto

* Rotate the cameras to face correct direction

* Remove noise and antiAliasing

* Add accelerometer and gyro

Positioned at [0,0,0.2] from torso space origin.

* Added Accelerometer & Gyro Placeholders

Added placeholders for the required parameters of both sensor objects. Current values are not correct and need to be edited.

* Accel Params Complete

The LUT table and the resolution parameters have been implemented accordingly to the ISS331dlh accelerometer.

* Gyroscope Params Filled

Completed the LUT list for the Gyroscope and the resolution value. Assumed sample rate was 400Hz in order to get the noise values.

* Added Lens and Focus Nodes to Cameras

Added nodes to camera objects for  convenience in future development.

* Added helpful comments and fixed Bug

Added comments describing the calculation Process of the noise values and assumptions made. In addition it was found that the first LUT values were incorrect and were corrected in this commit.

* Revert "Added helpful comments and fixed Bug"

This reverts commit 90847496dcddfe9841e96aa997d653684c27923b.

* Add messages we will be using

* Set up skeleton to begin handling messages

* Remove vs code folders and fix some formatting

* Add position sensor resolution

Sourced from the model specifications document

* WIP: read data

takes an actuator requests and does stuff

* It compiles!

* WIP: sending data

* Improved LUT

I have modified the LUT tables such that the RV's have a simulated 'DC gain' applied to them. This is advantageous as we work up from zero, as opposed to working up from -x to +x, as there will be no asymptote with the noise%

* Fixed Resolution Mismatch

The data calculated within the report did not match the data within the .PROTO file, hence it was updated.

* Update .gitignore

* Revert gitignore

* Update camera

* Add max focal distance

* Add reasonable noise and motionblur values

these are just guessed, they look sort of ok

* Adjust motion blur

and formatting

* Minimal noise

* It compiles??

* Formatting

* Fix up comments

* Believe I've merged 'main' into 'Sensors'

* Revert "Believe I've merged 'main' into 'Sensors'"

This reverts commit fc8fd9ac96cc810f6ec6aa2ca6ea4048e012e9ca.

* Add time info

* Use raw pointers  for things we dont have ownership over

* Check that messages exist before accessing them

* Fixes brace issue

* Moves to using switch statements instead of spicy casts

* Fixed ownership issue with device pointers

co-authored-by:Josephus Paye II <j.paye96@gmail.com>

* Fix the way we create the SensorMeasurements message

* Revert any nugus proto changes

not needed in this pr

* Added TODO List

* Fixed Accel LUT Units

* Update gyro and accelerometer LUT

* Remove unneeded texture

* Update NUgus.proto

* Fix up comment

* Add name field for gyro and accelerometer

* Give the sensors names

* Merge conflicts

wip

* Fixes cmake merge issue

* Add waitall flag to receiving data

I assume we want this

* Apply suggestions from code review

Co-authored-by: Alex Biddulph <c3124185@uon.edu.au>

* Change from if to #ndebug

I think this is what Alex wanted?

* Fix errors

and fix up throws

* Fix message typo

* Make timestep smaller

* Use @TrentHouliston's method for receiving messages

* Simplify message sending

* Formatting

* Removes the protobuf warnings

* Clang-tidy auto-apply-replacements

* Fixes warnings that I wasn't getting locally, but the buildkite machine had

Co-authored-by: Thomas O'Brien <thomas16441@gmail.com>
Co-authored-by: Benjamin Young <benjaminyoung7198@gmail.com>
Co-authored-by: Alex Biddulph <abiddulph@4tel.com.au>
Co-authored-by: Kip Hamiltons <p0nage@live.com.au>
Co-authored-by: Josephus Paye II <j.paye96@gmail.com>
Co-authored-by: Alex Biddulph <c3124185@uon.edu.au>